# src/main/java/com/prueba_tecnica/API_RESTFul/entity/Product.java
package com.prueba_tecnica.API_RESTFul.entity;


import jakarta.persistence.*;
import jakarta.validation.constraints.Min;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Entity
@Table(name = "product")
@Data
@NoArgsConstructor
@AllArgsConstructor
public class Product {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Integer id;

    @Column(name = "name")
    private String name;

    @Column(name = "description")
    private String description;

    @Column(name = "price")
    @Min(value = 1, message = "The price must be greater than 0")
    private float price;

    @Column(name = "category")
    private String category;


}